requirements.txt
langchain
langchain-google-genai
pypdf
python-docx
fastapi
uvicorn
chainlit
langgraph
pydantic
